requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
numpy>=1.23.0

# OCR and Document Processing
paddlepaddle>=2.4.0
paddleocr>=2.6.0

# Data Processing
pandas>=1.5.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Development Tools
jupyter>=1.0.0
ipykernel>=6.0.0

# Utilities
tqdm>=4.64.0
python-dateutil>=2.8.0

# LayoutLMv3 Specific
layoutlmv3>=0.1.0

# For testing
pytest>=7.0.0

# For model serving (optional)
fastapi>=0.85.0
uvicorn>=0.19.0
